Moved itoa from build.c to string.c since everybody seems to love it and copy it everywhere... :-/ Fix a 64-bit issue.
git-svn-id: http://svn.cri.ensmp.fr/svn/newgen/trunk@942 80f2d23d-f6f1-0310-8ea7-b0e6515bd809

+++ src/genC/list.c
@@ -631,7 +631,7 @@ list gen_bool_cons(bool b, list l)
   return gen_cons((void *) b, l);
 }
 
-list gen_int_cons(int i, list l)
+list gen_int_cons(intptr_t i, list l)
 {
   return gen_cons((void *) i, l);
 }
